feat(engine): implement next batch of keywords (#60) (#244)
* feat(engine): implement next batch of keywords — Annihilator, Undying, Persist, Afterlife, Exploit, Fabricate, Adapt, Modular (#60)

Wire 8 new keyword mechanics via generate_keyword_triggers/abilities in
card/mod.rs, matching Java CardFactoryUtil patterns. Add supporting
effect parameters: WithCountersType$ (ChangeZone ETB counters for
Undying/Persist), Amount$ (multi-sacrifice for Annihilator), Adapt$
(conditional counter gating), and defending_player propagation in
trigger handler for attack triggers.


* feat(engine): implement Ward, Crew, Embalm, Eternalize, Devotion keywords (#60)

Second batch of keyword implementations:
- Ward: BecomesTarget trigger that counters the targeting spell
- Crew: tapXType + Animate activated ability for vehicles
- Embalm/Eternalize: graveyard-activated CopyPermanent abilities
- Devotion: Count$Devotion.COLOR SVar resolver for mana symbol counting
- Add keyword_advanced2 preset deck for testing


* feat(engine): implement Morph, Megamorph, Bestow, Ninjutsu keywords (#60)

- Morph: cast face-down as 2/2 for {3}, turn face up for morph cost
- Megamorph: same as morph + adds +1/+1 counter on turn face up
- Bestow: alternative cast as Aura, falls off as creature when host leaves
- Ninjutsu: swap unblocked attacker from hand, enters tapped and attacking
- Frontend: MORPH/BESTOW badges, face-down card DTO masking
- Add preset deck (keyword_advanced3) for testing


* refactor(engine): improve maintainability of Morph/Bestow/Ninjutsu code

- Delete dead duplicate AlternativeCost enum (alternative_cost.rs was never imported)
- Add MORPH_GENERIC_COST and MORPH_PT constants to replace magic numbers
- Add AlternativeCost::is_morph() helper for Morph/Megamorph matching
- Add SpellAbility::param_is_true() helper to reduce repeated eq_ignore_ascii_case("True") pattern
- Use constants in game_action_util, magic_stack, and game_view_dto


* refactor(ui): extract card badge constants and reusable CardBadge component

- Add CARD_BADGES config to game.constants.ts (label + style per badge type)
- Add CARD_BANNER_CONTAINER and CARD_BANNER_TEXT to game.styles.ts
- Extract CardBadge sub-component to replace 4 copy-pasted banner blocks
- Use priority-based ternary chain (exerted > morph > bestow > transformed > token)
- Text-fallback badge also uses CARD_BADGES constants for colors/labels


* feat(ui): add search bar to deck selection dialog

Filter preset and user decks by name or description. The search input
is sticky at the top of the deck picker panel with a search icon.
Both sections show an empty-state message when no decks match.


* parity dash dep

* feat(engine): implement 10 new keywords + fix Fabricate/Modular/Ward (#60)

Fixes:
- Fabricate: player now chooses between +1/+1 counters or Servo tokens (SP$ Charm)
- Modular: death trigger targets artifact creature instead of self (SP$ Charm + ValidTgts$)
- Ward: opponent prompted to pay cost via UnlessCost$ (confirm_action)

New keywords:
- Outlast: activated ability (tap + mana → +1/+1 counter, sorcery speed)
- Unearth: activated from graveyard, grants Haste on ETB
- Exalted: Attacks trigger with Alone$ True, pumps attacker +1/+1
- Renown: combat damage trigger, places counters + sets is_renowned flag
- Flanking: AttackerBlockedByCreature trigger, blocker gets -1/-1
- Extort: optional SpellCast trigger, drains 1 life (LoseLife + GainLife chain)
- Bloodthirst: ETB counters gated on opponent life loss this turn
- Riot: ETB choice via SP$ Charm (counter or haste)
- Unleash: ETB +1/+1 counter + can't block static in layer system

Infrastructure: Attacks trigger alone field, num_attackers in RunParams,
AttackerBlockedByCreature fired in phase_handler, TriggeredAttacker/Blocker
propagation in handler, Renown$/Bloodthirst$ gating in counters_put_effect.


* fix missing stuff

* fix: equip

* feat(ui): render equipment and aura attachments stacked on host cards

Add attachedTo/attachmentIds fields to CardDto and TypeScript Card interface.
Both BattlefieldZone and FreeBattlefield now group attached cards with their
host, rendering them stacked with a vertical offset. Attached cards are filtered
from the grid layout and each card remains independently hoverable.


* feat(ui): add P/T color-coding and phased-out dimming

P/T overlay is now green when buffed above base, red when debuffed,
amber when mixed. Phased-out cards render at 30% opacity with grayscale.
Adds basePower/baseToughness fields to CardDto for comparison.


* Fix unused import, variables, and unreachable statement warnings


* Fix Persist/Undying LKI: emit zone trigger before move_card in SBA

When a creature dies via state-based actions, the ChangesZone trigger
must be matched while the card still has its battlefield state (counters,
keywords). Previously the trigger was emitted after move_card which
clears counters, causing Persist to always fire (counters_EQ0_M1M1
trivially true on a cleared card). Now we emit + flush before the move
so the counter check uses Last Known Information.


* fix fabricate

* fix: mana creat

* Route ManaReflected color choice through agent for RNG parity

Java's DeterministicController.chooseColor() uses pickIndex(colorList.size(), rng)
when resolving ManaReflected abilities (e.g. Incubation Druid). The Rust engine was
using a hardcoded deterministic heuristic (pick color with least mana in pool) that
consumed no RNG, causing parity divergence. Now routes through the agent's
choose_color() method with colors sorted in WUBRG order to match Java's ColorSet
iteration.


* Implement Exploit keyword: optional sacrifice with Exploited trigger

- sacrifice_effect.rs: Add do_sacrifice helper with Exploit$ True detection.
  When Optional$ True with SacValid != Self, use choose_cards_for_effect(min=0)
  matching Java's choosePermanentsToSacrifice(min=0, max=amount). Fires
  TriggerType::Exploited after sacrifice when Exploit$ True is set.
- event/mod.rs: Add Exploited variant to TriggerType and exploited_card to RunParams
- trigger/trigger.rs: Add TriggerMode::Exploited with ValidCard/ValidSource,
  perform_test matching, and parser for Mode$ Exploited
- card/mod.rs: Add Exploit$ True to TrigExploit SVar
- handler.rs, delayed_trigger_effect.rs, static_ability_*.rs: exhaustive match arms


* fix: keyword_advanced

* add new reg

* fix some parity

* fix reg

* refix deploy sh

* skill md

* fix: parity!!!!!

* fix: parity!!!!!

* ui

* fix: parity again

* fix: more reg

* feat: agentic harness check

* feat: auto harness check

* flacky test

---------

+++ forge-engine/crates/forge-parity/src/parity_card_map.rs
@@ -39,6 +39,7 @@ impl ParityCardMap {
         self.by_card
             .get(&cid)
             .copied()
-            .unwrap_or(u32::MAX.saturating_sub(cid.0))
+            // Use i32::MAX to match Java's Integer.MAX_VALUE fallback for tokens.
+            .unwrap_or((i32::MAX as u32).saturating_sub(cid.0))
     }
 }
